voice-agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
redis==5.0.7
websockets==12.0
httpx==0.27.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy==1.26.4
pydantic-settings==2.3.4
pytest==8.2.2
pytest-asyncio==0.23.7
